# Optimize unserializing arrays with apc.serializer=default