The pointer data apcu sees is usually aligned to 16 bytes on 64-bit systems using emalloc. It is guaranteed to be aligned to a power of at least ZEND_MM_ALIGNMENT_LOG2.
Preserve the original bits just in case, similar to opcache's zend_rotr3.

Benchmarking this for repeatedly unserializing large arrays with small strings and apc.serializer=default, this showed a moderate performance improvement. (note that apcu_fetch returns a different array on every call due to the possibility of cache evictions) (reducing cpu frequency to 2.5GHz to be the same for both tests)

For an array with 65536 small string keys and values and `apc.serializer=default`:
- Repeated apcu_fetch went from 3.5 seconds -> 3.0 seconds
- For an array with 4096 small string keys and values: Repeated apcu_fetch went from 1.2 seconds -> 0.9 seconds

For #323

After this change:

```
==> results_optimized_apc_default.txt <==
bench_fetch_large_config.php Results for php 8.2.0-dev debug=0 with opcache enabled=true, apc.serializer=default

bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.014
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=0.912
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=0.889
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.277
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.075
bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.005
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=0.909
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=0.916
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.275
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.073
```

Before this change (took longer):
```
bench_fetch_large_config.php Results for php 8.2.0-dev debug=0 with opcache enabled=true, apc.serializer=default

bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.513
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=1.233
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=1.094
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.259
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.034
bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.508
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=1.209
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=1.091
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.284
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.036
```

<details>

<summary>Benchmark source code (click to expand)</summary>

```php
<?php // bench_fetch_large_config.php

printf(
    "%s Results for php %s debug=%s with opcache enabled=%s, apc.serializer=%s\n\n",
    basename(__FILE__),
    PHP_VERSION,
    json_encode(PHP_DEBUG),
    json_encode(function_exists('opcache_get_status') && (opcache_get_status(false)['opcache_enabled'] ?? false)),
    ini_get('apc.serializer')
);

function bench_apcu_fetch_large_config(int $numValues, int $repetitions) {
    $values = [];
    srand(1234);
    for ($i = 0; $i < $numValues; $i++) {
        $values["key" . rand(0, PHP_INT_MAX)] = "value" . rand(0, PHP_INT_MAX);
    }
    if (!apcu_store('config_key', $values)) {
        throw new Exception("Failed to set config_key");
    }
    $expected = count($values);
    $start = hrtime(true);
    for ($i = 0; $i < $repetitions; $i++) {
        $v = apcu_fetch('config_key');
        if (count($v) !== $expected) {
            throw new RuntimeException("Failed to load config");
        }
    }
    $end = hrtime(true);
    printf("%s: numValues=%10d repetitions=%10d elapsed=%.3f\n", __FUNCTION__, $numValues, $repetitions, ($end - $start)/1e9);
}

$settings = [
    [2**16, 2**8],
    [2**12, 2**12],
    [2**8,  2**16],
    [2**3,  2**21],
    [1,     2**22],
];
for ($i = 0; $i < 2; $i++) {
    foreach ($settings as [$n, $repetitions]) {
        bench_apcu_fetch_large_config($n, $repetitions);
    }
}

```

</details>