@TysonAndre 2022-10-23
Since this improves the performance of `apc.serializer=default` for the use case of large configs,
I ran the benchmarks of the other serializers to compare this against.

PHP

```
bench_fetch_large_config.php Results for php 8.2.0-dev debug=0 with opcache enabled=true, apc.serializer=php

bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=2.598
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=1.435
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=1.383
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.915
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.454
bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=2.478
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=1.463
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=1.384
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.926
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.448
```

igbinary. The current v2 protocol adds all unserialized strings ids to a table to look up later even if it doesn't need the string.
So it's performing worse for this benchmark, which is entirely distinct strings
(and php unserialize() has no string reuse)
That's likely why the serialization performance is slower for large numValues (I'm a lead of igbinary)

```
bench_fetch_large_config.php Results for php 8.2.0-dev debug=0 with opcache enabled=true, apc.serializer=igbinary

bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.007
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=0.927
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=0.891
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.260
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.051
bench_apcu_fetch_large_config: numValues=     65536 repetitions=       256 elapsed=3.009
bench_apcu_fetch_large_config: numValues=      4096 repetitions=      4096 elapsed=0.915
bench_apcu_fetch_large_config: numValues=       256 repetitions=     65536 elapsed=0.907
bench_apcu_fetch_large_config: numValues=         8 repetitions=   2097152 elapsed=1.282
bench_apcu_fetch_large_config: numValues=         1 repetitions=   4194304 elapsed=1.032
```